Created and implemented a custom hook for using the Escape key
